requirements.txt
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.2.0
decorator==5.1.1
et-xmlfile==1.1.0
exceptiongroup==1.1.2
ffmpeg==1.4
ffmpy==0.3.1
gevent==23.7.0
greenlet==2.0.2
h11==0.14.0
idna==2.10
Jinja2==3.0.3
kafka-python==2.0.2
MarkupSafe==2.1.3
numpy==1.24.4
opencv-python==4.8.0.74
openpyxl==3.1.2
outcome==1.2.0
py==1.11.0
pycparser==2.21
PyMySQL==1.1.0
PySocks==1.7.1
PyYAML==5.3.1
requests==2.24.0
retry==0.9.2
selenium==4.10.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.2
trio-websocket==0.10.3
unittestreport==1.5.1
urllib3==1.25.11
websocket==0.2.1
wsproto==1.2.0
zope.event==5.0
zope.interface==6.0
